+++ src/iqm/iqm_client/iqm_client.py
@@ -23,7 +23,6 @@ import itertools
 import json
 import os
 import platform
-from posixpath import join
 import time
 from typing import Any, Callable, Optional
 from uuid import UUID
@@ -31,6 +30,7 @@ import warnings
 
 import requests
 
+from iqm.iqm_client.api import APIConfig, APIEndpoint, APIVariant
 from iqm.iqm_client.authentication import TokenManager
 from iqm.iqm_client.errors import (
     APITimeoutError,
@@ -48,6 +48,7 @@ from iqm.iqm_client.models import (
     CircuitCompilationOptions,
     DynamicQuantumArchitecture,
     Instruction,
+    MoveGateValidationMode,
     QuantumArchitecture,
     QuantumArchitectureSpecification,
     RunRequest,
@@ -80,10 +81,12 @@ class IQMClient:
             If ``auth_server_url`` is given also ``username`` and ``password`` must be given.
         username: Username to log in to authentication server.
         password: Password to log in to authentication server.
+        api_variant: API variant to use. Default is ``APIVariant.COCOS``.
+            Conviguable also by environment variable ``IQM_CLIENT_API_VARIANT``.
 
     Alternatively, the user authentication related keyword arguments can also be given in
-    environment variables ``IQM_TOKEN``, ``IQM_TOKENS_FILE``, ``IQM_AUTH_SERVER``,
-    ``IQM_AUTH_USERNAME`` and ``IQM_AUTH_PASSWORD``. All parameters must be given either
+    environment variables :envvar:`IQM_TOKEN`, :envvar:`IQM_TOKENS_FILE`, :envvar:`IQM_AUTH_SERVER`,
+    :envvar:`IQM_AUTH_USERNAME` and :envvar:`IQM_AUTH_PASSWORD`. All parameters must be given either
     as keyword arguments or as environment variables. Same combination restrictions apply
     for values given as environment variables as for keyword arguments.
     """
@@ -98,6 +101,7 @@ class IQMClient:
         auth_server_url: Optional[str] = None,
         username: Optional[str] = None,
         password: Optional[str] = None,
+        api_variant: Optional[APIVariant] = None,
     ):
         if not url.startswith(('http:', 'https:')):
             raise ClientConfigurationError(f'The URL schema has to be http or https. Incorrect schema in URL: {url}')
@@ -108,7 +112,6 @@ class IQMClient:
             username,
             password,
         )
-        self._base_url = url
         self._signature = f'{platform.platform(terse=True)}'
         self._signature += f', python {platform.python_version()}'
         self._signature += f', iqm-client {version("iqm-client")}'
@@ -116,6 +119,11 @@ class IQMClient:
             self._signature += f', {client_signature}'
         self._architecture: QuantumArchitectureSpecification | None = None
         self._dynamic_architectures: dict[UUID, DynamicQuantumArchitecture] = {}
+
+        if api_variant is None:
+            env_var = os.environ.get('IQM_CLIENT_API_VARIANT')
+            api_variant = APIVariant(env_var) if env_var else APIVariant.V1
+        self._api = APIConfig(api_variant, url)
 
     def __del__(self):
         try:
@@ -152,15 +160,15 @@ class IQMClient:
         """Submits a batch of quantum circuits for execution on a quantum computer.
 
         Args:
-            circuits: list of circuits to be executed
+            circuits: Circuits to be executed.
             qubit_mapping: Mapping of logical qubit names to physical qubit names.
                 Can be set to ``None`` if all ``circuits`` already use physical qubit names.
                 Note that the ``qubit_mapping`` is used for all ``circuits``.
             custom_settings: Custom settings to override default settings and calibration data.
                 Note: This field should always be ``None`` in normal use.
             calibration_set_id: ID of the calibration set to use, or ``None`` to use the latest one
-            shots: number of times ``circuits`` are executed, value must be greater than zero
-            options: Various discrete options for compiling quantum circuits to pulse schedules.
+            shots: Number of times ``circuits`` are executed. Must be greater than zero
+            options: Various discrete options for compiling quantum circuits to instruction schedules.
         Returns:
             ID for the created job. This ID is needed to query the job status and the execution results.
         """
@@ -193,19 +201,15 @@ class IQMClient:
         submitting it for execution.
 
         Args:
-            circuits: list of circuits to be executed
+            circuits: Circuits to be executed.
             qubit_mapping: Mapping of logical qubit names to physical qubit names.
                 Can be set to ``None`` if all ``circuits`` already use physical qubit names.
                 Note that the ``qubit_mapping`` is used for all ``circuits``.
             custom_settings: Custom settings to override default settings and calibration data.
                 Note: This field should always be ``None`` in normal use.
             calibration_set_id: ID of the calibration set to use, or ``None`` to use the latest one
-            shots: number of times ``circuits`` are executed, value must be greater than zero
-            max_circuit_duration_over_t2: Circuits are disqualified on the server if they are longer than this ratio
-                of the T2 time of the qubits. Setting this value to ``0.0`` turns off circuit duration checking.
-                The default value ``None`` instructs server to use server's default value in the checking.
-            heralding_mode: Heralding mode to use during the execution.
-
+            shots: Number of times ``circuits`` are executed. Must be greater than zero.
+            options: Various discrete options for compiling quantum circuits to instruction schedules.
         Returns:
             RunRequest that would be submitted by equivalent call to :meth:`submit_circuits`.
         """
@@ -229,7 +233,9 @@ class IQMClient:
         serialized_qubit_mapping = serialize_qubit_mapping(qubit_mapping) if qubit_mapping else None
 
         # validate the circuit against the quantum architecture of the server
-        self._validate_circuit_instructions(architecture, circuits, qubit_mapping)
+        self._validate_circuit_instructions(
+            architecture, circuits, qubit_mapping, validate_moves=options.move_gate_validation
+        )
 
         return RunRequest(
             qubit_mapping=serialized_qubit_mapping,
@@ -250,7 +256,6 @@ class IQMClient:
 
         Args:
             run_request: the run request to be submitted for execution.
-
         Returns:
             ID for the created job. This ID is needed to query the job status and the execution results.
         """
@@ -271,7 +276,7 @@ class IQMClient:
 
         result = self._retry_request_on_error(
             lambda: requests.post(
-                join(self._base_url, 'jobs'),
+                self._api.url(APIEndpoint.SUBMIT_JOB),
                 json=json.loads(run_request.model_dump_json(exclude_none=True)),
                 headers=headers,
                 timeout=REQUESTS_TIMEOUT,
@@ -301,14 +306,14 @@ class IQMClient:
         """Validates the given qubit mapping, if defined.
 
         Args:
-          architecture: the quantum architecture to check against
-          circuits: list of circuits to be checked
+          architecture: Quantum architecture to check against.
+          circuits: Circuits to be checked.
           qubit_mapping: Mapping of logical qubit names to physical qubit names.
               Can be set to ``None`` if all ``circuits`` already use physical qubit names.
               Note that the ``qubit_mapping`` is used for all ``circuits``.
 
         Raises:
-            CircuitExecutionError: IQM server specific exceptions
+            CircuitValidationError: There was something wrong with ``circuits``.
         """
         if qubit_mapping is None:
             return
@@ -316,13 +321,13 @@ class IQMClient:
         # check if qubit mapping is injective
         target_qubits = set(qubit_mapping.values())
         if not len(target_qubits) == len(qubit_mapping):
-            raise ValueError('Multiple logical qubits map to the same physical qubit.')
+            raise CircuitValidationError('Multiple logical qubits map to the same physical qubit.')
 
         # check if qubit mapping covers all qubits in the circuits
         for i, circuit in enumerate(circuits):
             diff = circuit.all_qubits() - set(qubit_mapping)
             if diff:
-                raise ValueError(
+                raise CircuitValidationError(
                     f"The qubits {diff} in circuit '{circuit.name}' at index {i} "
                     f'are not found in the provided qubit mapping.'
                 )
@@ -330,14 +335,15 @@ class IQMClient:
         # check that each mapped qubit is defined in the quantum architecture
         for _logical, physical in qubit_mapping.items():
             if physical not in architecture.qubits:
-                raise CircuitExecutionError(f'Qubit {physical} not present in quantum architecture')
+                raise CircuitValidationError(f'Qubit {physical} not present in quantum architecture')
 
     @staticmethod
     def _validate_circuit_instructions(
         architecture: QuantumArchitectureSpecification,
         circuits: CircuitBatch,
         qubit_mapping: Optional[dict[str, str]] = None,
-    ):
+        validate_moves: MoveGateValidationMode = MoveGateValidationMode.STRICT,
+    ) -> None:
         """Validate the given circuits against the given quantum architecture.
 
         Args:
@@ -346,14 +352,22 @@ class IQMClient:
           qubit_mapping: Mapping of logical qubit names to physical qubit names.
               Can be set to ``None`` if all ``circuits`` already use physical qubit names.
               Note that the ``qubit_mapping`` is used for all ``circuits``.
+          validate_moves: determines how MOVE gate validation works
 
         Raises:
             CircuitExecutionError: validation failed
         """
-        for circuit in circuits:
-            IQMClient._validate_circuit_moves(architecture, circuit, qubit_mapping)
+        for index, circuit in enumerate(circuits):
+            measurement_keys: set[str] = set()
             for instr in circuit.instructions:
                 IQMClient._validate_instruction(architecture, instr, qubit_mapping)
+                # check measurement key uniqueness
+                if instr.name in {'measure', 'measurement'}:
+                    key = instr.args['key']
+                    if key in measurement_keys:
+                        raise CircuitValidationError(f'Circuit {index}: {instr!r} has a non-unique measurement key.')
+                    measurement_keys.add(key)
+            IQMClient._validate_circuit_moves(architecture, circuit, qubit_mapping, validate_moves=validate_moves)
 
     @staticmethod
     def _validate_instruction(
@@ -361,7 +375,7 @@ class IQMClient:
         instruction: Instruction,
         qubit_mapping: Optional[dict[str, str]] = None,
     ) -> None:
-        """Validate that the instruction targets a valid qubit locus in the given quantum architecture.
+        """Check that the instruction targets a valid qubit locus in the given quantum architecture.
 
         Args:
           architecture: quantum architecture to check against
@@ -373,7 +387,7 @@ class IQMClient:
             CircuitExecutionError: validation failed
         """
         if instruction.name not in architecture.operations:
-            raise CircuitExecutionError(
+            raise CircuitValidationError(
                 f"Instruction '{instruction.name}' is not supported by the quantum architecture."
             )
         allowed_loci = architecture.operations[instruction.name]
@@ -396,7 +410,7 @@ class IQMClient:
             allowed_qubits = set(q for locus in allowed_loci for q in locus)
             for q, mapped_q in zip(instruction.qubits, qubits):
                 if mapped_q not in allowed_qubits:
-                    raise CircuitExecutionError(
+                    raise CircuitValidationError(
                         f'Qubit {q} = {mapped_q} is not allowed as locus for {instruction.name}'
                         if qubit_mapping
                         else f'Qubit {q} is not allowed as locus for {instruction.name}'
@@ -410,7 +424,7 @@ class IQMClient:
             else allowed_loci
         )
         if qubits not in all_loci:
-            raise CircuitExecutionError(
+            raise CircuitValidationError(
                 f'{instruction.qubits} = {tuple(qubits)} not allowed as locus for {instruction.name}'
                 if qubit_mapping
                 else f'{instruction.qubits} not allowed as locus for {instruction.name}'
@@ -418,66 +432,204 @@ class IQMClient:
 
     @staticmethod
     def _validate_circuit_moves(
-        architecture: QuantumArchitectureSpecification, circuit: Circuit, qubit_mapping: Optional[dict[str, str]] = None
+        architecture: QuantumArchitectureSpecification,
+        circuit: Circuit,
+        qubit_mapping: Optional[dict[str, str]] = None,
+        validate_moves: MoveGateValidationMode = MoveGateValidationMode.STRICT,
     ) -> None:
-        """Validates that the MOVE gates in the circuit are not exciting the resonator.
+        """Raises an error if the MOVE gates in the circuit are not valid in the given architecture.
 
         Args:
             architecture: Quantum architecture to check against.
             circuit: Quantum circuit to validate.
             qubit_mapping: Mapping of logical qubit names to physical qubit names.
                 Can be set to ``None`` if the ``circuit`` already uses physical qubit names.
+            validate_moves: Option for bypassing full or partial MOVE gate validation.
         Raises:
             CircuitExecutionError: validation failed
         """
+        # pylint: disable=too-many-branches
+        if validate_moves == MoveGateValidationMode.NONE:
+            return
         move_gate = 'move'
         # Check if MOVE gates are allowed on this architecture
         if move_gate not in architecture.operations:
             if any(i.name == move_gate for i in circuit.instructions):
-                raise CircuitExecutionError('MOVE instruction is not supported by the given device architecture.')
+                raise CircuitValidationError('MOVE instruction is not supported by the given device architecture.')
             return
-        # Track the location of the resonator state
-        # TODO use architecture.computational_resonators when available instead of using COMP_R.
+
+        # some gates are allowed in MOVE sandwiches
+        allowed_gates = {'barrier'}
+        if validate_moves == MoveGateValidationMode.ALLOW_PRX:
+            allowed_gates.add('prx')
+
+        # TODO use architecture.computational_resonators when available instead of relying on COMP_R prefix.
+        all_resonators = {q for q in architecture.qubits if q.startswith('COMP_R')}
+        all_qubits = set(architecture.qubits) - all_resonators
         if qubit_mapping:
             reverse_mapping = {phys: log for log, phys in qubit_mapping.items()}
-            resonator_state_loc = {  # Resonator: Location
-                reverse_mapping[q]: reverse_mapping[q] for q in architecture.qubits if q.startswith('COMP_R')
-            }
-        else:
-            resonator_state_loc = {q: q for q in architecture.qubits if q.startswith('COMP_R')}  # Resonator: Location
-        for instr in circuit.instructions:
-            # If any of the gate arguments are holding the |0> state of the resonator, we need to check the instruction
-            if any(qb in resonator_state_loc.values() for qb in instr.qubits):
-                # We are using a qubit or resonator state that is currently in the resonator.
-                if instr.name == move_gate:
-                    qb, res = instr.qubits
-                    # Check if qb is a qubit and res a resonator.
-                    if res not in resonator_state_loc or qb in resonator_state_loc:
-                        raise CircuitExecutionError(
-                            f'MOVE instruction only allowed between qubit and resonator, not {instr.qubits}.'
-                        )
-                    if resonator_state_loc[res] not in [qb, res]:
-                        raise CircuitExecutionError(
-                            f'MOVE instruction between {instr.qubits} while qubit state is in another resonator.'
-                        )
-                    # Update the resonator state location
-                    resonator_state_loc[res] = qb if resonator_state_loc[res] == res else res
-                elif instr.name not in ['barrier'] and any(
-                    qb in resonator_state_loc.values() and (qb, qb) not in resonator_state_loc.items()
-                    for qb in instr.qubits
-                ):
-                    # The instruction is using a qubit that is holding a resonator state but it is not the resonator
-                    raise CircuitExecutionError(
-                        f'Instruction {instr.name} on {instr.qubits} while they hold a resonator state. \
-                            Qubit states currently moved to resonator(s): {resonator_state_loc}.'
+            all_resonators = {reverse_mapping[q] if q in reverse_mapping else q for q in all_resonators}
+            all_qubits = {reverse_mapping[q] if q in reverse_mapping else q for q in all_qubits}
+
+        # Mapping from resonator to the qubit whose state it holds. Resonators not in the map hold no qubit state.
+        resonator_occupations: dict[str, str] = {}
+        # Qubits whose states are currently moved to a resonator
+        moved_qubits: set[str] = set()
+
+        for inst in circuit.instructions:
+            if inst.name == 'move':
+                qubit, resonator = inst.qubits
+                if not (qubit in all_qubits and resonator in all_resonators):
+                    raise CircuitValidationError(
+                        f'MOVE instructions are only allowed between qubit and resonator, not {inst.qubits}.'
                     )
-            elif instr.name == move_gate:
-                raise CircuitExecutionError(
-                    f'MOVE instruction between {instr.qubits} while neither holds a resonator state, and neither \
-                    qubits are in a resonator.'
-                )
-        if any(res != qb for res, qb in resonator_state_loc.items()):
-            raise CircuitExecutionError('Circuit ends while qubit state still in the resonator.')
+
+                if (resonator_qubit := resonator_occupations.get(resonator)) is None:
+                    # Beginning MOVE: check that the qubit hasn't been moved to another resonator
+                    if qubit in moved_qubits:
+                        raise CircuitValidationError(
+                            f'MOVE instruction {inst.qubits}: state of {qubit} is '
+                            f'in another resonator: {resonator_occupations}.'
+                        )
+                    resonator_occupations[resonator] = qubit
+                    moved_qubits.add(qubit)
+                else:
+                    # Ending MOVE: check that the qubit matches to the qubit that was moved to the resonator
+                    if resonator_qubit != qubit:
+                        raise CircuitValidationError(
+                            f'MOVE instruction {inst.qubits} to an already occupied resonator: {resonator_occupations}.'
+                        )
+                    del resonator_occupations[resonator]
+                    moved_qubits.remove(qubit)
+            elif moved_qubits:
+                # Validate that moved qubits are not used during MOVE operations
+                if inst.name not in allowed_gates:
+                    if overlap := set(inst.qubits) & moved_qubits:
+                        raise CircuitValidationError(
+                            f'Instruction {inst.name} acts on {inst.qubits} while the state(s) of {overlap} '
+                            f'are in a resonator. Current resonator occupation: {resonator_occupations}.'
+                        )
+
+        # Finally validate that all moves have been ended before the circuit ends
+        if resonator_occupations:
+            raise CircuitValidationError(
+                f'Circuit ends while qubit state(s) are still in a resonator: {resonator_occupations}.'
+            )
+
+    def _get_run_v1(self, job_id: UUID, timeout_secs: float = REQUESTS_TIMEOUT) -> RunResult:
+        """
+        V1 API (Cocos circuits execution and Resonance) has an inefficient `GET /jobs/<id>` endpoint
+        that returns the full job status, including the result and the original request, in a single call.
+        """
+        result = self._retry_request_on_error(
+            lambda: requests.get(
+                self._api.url(APIEndpoint.GET_JOB_RESULT, str(job_id)),
+                headers=self._default_headers(),
+                timeout=timeout_secs,
+            )
+        )
+        result.raise_for_status()
+        return RunResult.from_dict(result.json())
+
+    def _get_run_v2(self, job_id: UUID, timeout_secs: float = REQUESTS_TIMEOUT) -> RunResult:
+        """
+        V2 API (Station-based circuits execution) has granular endpoints for job status and result.
+        """
+        status_response = requests.get(
+            self._api.url(APIEndpoint.GET_JOB_STATUS, str(job_id)),
+            headers=self._default_headers(),
+            timeout=timeout_secs,
+        )
+        status_response.raise_for_status()
+        status = status_response.json()
+        if Status(status['status']) not in [Status.READY, Status.ABORTED, Status.DELETED, Status.FAILED]:
+            return RunResult.from_dict(
+                {
+                    'measurements': [],
+                    'status': status['status'],
+                    'message': '',
+                    'metadata': {
+                        'calibration_set_id': None,
+                        'circuits_batch': [],
+                        'parameters': None,
+                        'timestamps': {},
+                    },
+                }
+            )
+
+        result = self._retry_request_on_error(
+            lambda: requests.get(
+                self._api.url(APIEndpoint.GET_JOB_RESULT, str(job_id)),
+                headers=self._default_headers(),
+                timeout=timeout_secs,
+            )
+        )
+        if result.status_code != 404:
+            result.raise_for_status()
+        measurements = [] if result.status_code == 404 else result.json()
+        request_parameters = (
+            {}
+            if result.status_code == 404
+            else requests.get(
+                self._api.url(APIEndpoint.GET_JOB_REQUEST_PARAMETERS, str(job_id)),
+                headers=self._default_headers(),
+                timeout=timeout_secs,
+            ).json()
+        )
+        calibration_set_id = (
+            None
+            if result.status_code == 404
+            else requests.get(
+                self._api.url(APIEndpoint.GET_JOB_CALIBRATION_SET_ID, str(job_id)),
+                headers=self._default_headers(),
+                timeout=timeout_secs,
+            ).json()
+        )
+        circuits_batch = (
+            []
+            if result.status_code == 404
+            else requests.get(
+                self._api.url(APIEndpoint.GET_JOB_CIRCUITS_BATCH, str(job_id)),
+                headers=self._default_headers(),
+                timeout=timeout_secs,
+            ).json()
+        )
+        timeline = (
+            []
+            if result.status_code == 404
+            else requests.get(
+                self._api.url(APIEndpoint.GET_JOB_TIMELINE, str(job_id)),
+                headers=self._default_headers(),
+                timeout=timeout_secs,
+            ).json()
+        )
+        error_message_response = requests.get(
+            self._api.url(APIEndpoint.GET_JOB_ERROR_LOG, str(job_id)),
+            headers=self._default_headers(),
+            timeout=timeout_secs,
+        )
+        error_message = error_message_response.text if error_message_response.status_code == 200 else None
+        return RunResult.from_dict(
+            {
+                'measurements': measurements,
+                'status': status['status'],
+                'message': error_message,
+                'metadata': {
+                    '`': calibration_set_id,
+                    'circuits_batch': circuits_batch,
+                    'parameters': None
+                    if result.status_code == 404
+                    else {
+                        'shots': request_parameters['shots'],
+                        'max_circuit_duration_over_t2': request_parameters['max_circuit_duration_over_t2'],
+                        'heralding_mode': request_parameters['heralding_mode'],
+                        'move_validation_mode': request_parameters['move_validation_mode'],
+                        'move_gate_frame_tracking_mode': request_parameters['move_gate_frame_tracking_mode'],
+                    },
+                    'timestamps': {datapoint['stage']: datapoint['timestamp'] for datapoint in timeline},
+                },
+            }
+        )
 
     def get_run(self, job_id: UUID, *, timeout_secs: float = REQUESTS_TIMEOUT) -> RunResult:
         """Query the status and results of a submitted job.
@@ -493,19 +645,13 @@ class IQMClient:
             CircuitExecutionError: IQM server specific exceptions
             HTTPException: HTTP exceptions
         """
-        result = self._retry_request_on_error(
-            lambda: requests.get(
-                join(self._base_url, 'jobs', str(job_id)),
-                headers=self._default_headers(),
-                timeout=timeout_secs,
-            )
-        )
-
-        result.raise_for_status()
         try:
-            run_result = RunResult.from_dict(result.json())
+            if self._api.variant == APIVariant.V2:
+                run_result = self._get_run_v2(job_id, timeout_secs)
+            else:
+                run_result = self._get_run_v1(job_id, timeout_secs)
         except (json.decoder.JSONDecodeError, KeyError) as e:
-            raise CircuitExecutionError(f'Invalid response: {result.text}, {e}') from e
+            raise CircuitExecutionError(f'Invalid response: {e}') from e
 
         if run_result.warnings:
             for warning in run_result.warnings:
@@ -530,7 +676,7 @@ class IQMClient:
         """
         result = self._retry_request_on_error(
             lambda: requests.get(
-                join(self._base_url, 'jobs', str(job_id), 'status'),
+                self._api.url(APIEndpoint.GET_JOB_STATUS, str(job_id)),
                 headers=self._default_headers(),
                 timeout=timeout_secs,
             )
@@ -563,7 +709,7 @@ class IQMClient:
         start_time = datetime.now()
         while (datetime.now() - start_time).total_seconds() < timeout_secs:
             status = self.get_run_status(job_id).status
-            if status != Status.PENDING_COMPILATION:
+            if status not in (Status.PENDING_COMPILATION, Status.ACCEPTED, Status.PROCESSING, Status.RECEIVED):
                 return self.get_run(job_id)
             time.sleep(SECONDS_BETWEEN_CALLS)
         raise APITimeoutError(f"The job compilation didn't finish in {timeout_secs} seconds.")
@@ -587,7 +733,13 @@ class IQMClient:
         start_time = datetime.now()
         while (datetime.now() - start_time).total_seconds() < timeout_secs:
             status = self.get_run_status(job_id).status
-            if status not in (Status.PENDING_COMPILATION, Status.PENDING_EXECUTION):
+            if status not in (
+                Status.PENDING_COMPILATION,
+                Status.PENDING_EXECUTION,
+                Status.ACCEPTED,
+                Status.PROCESSING,
+                Status.RECEIVED,
+            ):
                 return self.get_run(job_id)
             time.sleep(SECONDS_BETWEEN_CALLS)
         raise APITimeoutError(f"The job didn't finish in {timeout_secs} seconds.")
@@ -604,7 +756,7 @@ class IQMClient:
             JobAbortionError: if aborting the job failed
         """
         result = requests.post(
-            join(self._base_url, 'jobs', str(job_id), 'abort'),
+            self._api.url(APIEndpoint.ABORT_JOB, str(job_id)),
             headers=self._default_headers(),
             timeout=timeout_secs,
         )
@@ -630,7 +782,7 @@ class IQMClient:
             return self._architecture
 
         result = requests.get(
-            join(self._base_url, 'quantum-architecture'),
+            self._api.url(APIEndpoint.QUANTUM_ARCHITECTURE),
             headers=self._default_headers(),
             timeout=timeout_secs,
         )
@@ -677,7 +829,7 @@ class IQMClient:
             calibration_set_id_str = str(calibration_set_id)
 
         result = requests.get(
-            join(self._base_url, f'api/v1/calibration/{calibration_set_id_str}/gates'),
+            self._api.url(APIEndpoint.CALIBRATED_GATES, calibration_set_id_str),
             headers=self._default_headers(),
             timeout=timeout_secs,
         )