+++ CHANGELOG.rst
@@ -10,6 +10,35 @@ Version 19.0
 * Remove the deprecated native operations names ``phased_rx`` and ``measurement``.
   `#136 <https://github.com/iqm-finland/iqm-client/pull/136>`_
 
+Version 18.8
+============
+
+* Fix MOVE gate validation for qubit mappings containing only some of the architecture qubits `#137 <https://github.com/iqm-finland/iqm-client/pull/137>`_
+
+Version 18.7
+============
+
+* Fix list of endpoints supported by the V1 API. `#138 <https://github.com/iqm-finland/iqm-client/pull/138>`_
+
+Version 18.6
+============
+
+* Add IQM Server API versioning support. `#135 <https://github.com/iqm-finland/iqm-client/pull/135>`_
+
+Version 18.5
+============
+
+* Added ``isort`` formatting to the tox configuration, so it is automatically run when running
+  ``tox -e format``. `#130 <https://github.com/iqm-finland/iqm-client/pull/130>`_
+* Bugfix: Fix the issue where the :class:`CircuitCompilationOptions` was not used in local circuit
+  validation when using the :meth:`submit_circuit` method. Improved testing to catch the bug.
+  `#130 <https://github.com/iqm-finland/iqm-client/pull/130>`_
+* Bugfix: MOVE gate validation now also works with more than one resonator. `#130 <https://github.com/iqm-finland/iqm-client/pull/130>`_
+* More specific validation and transpilation errors. `#130 <https://github.com/iqm-finland/iqm-client/pull/130>`_
+* Docs updated: mid-circuit measurements are allowed on stations with ``cocos >= 30.2``. `#130 <https://github.com/iqm-finland/iqm-client/pull/130>`_
+* Integration guide updated. `#130 <https://github.com/iqm-finland/iqm-client/pull/130>`_
+* Circuit validation: All measurement keys must be unique. `#130 <https://github.com/iqm-finland/iqm-client/pull/130>`_
+
 Version 18.4
 ============
 
@@ -35,11 +64,13 @@ Version 18.0
 
 * Added the naive MOVE transpilation method for unified transpilation behavior for different external APIs. `#124 <https://github.com/iqm-finland/iqm-client/pull/124>`_
 * Added class for compilation options :class:`CircuitCompilationOptions` to allow for more fine-grained control over the compilation process. (breaking change)
-    * :meth:`IQMClient.submit_circuit` now takes a :class:`CircuitCompilationOptions` parameter instead of ``max_circuit_duration_over_t2`` and ``heralding_mode``.
-    * Moved the existing ``max_circuit_duration_over_t2`` parameter to :class:`CircuitCompilationOptions`.
-    * Moved the existing ``heralding_mode`` parameter to :class:`CircuitCompilationOptions`.
-    * Introduced new option ``move_gate_validation`` to turn off MOVE gate validation during compilation (ADVANCED).
-    * Introduced new option ``move_gate_frame_tracking`` to turn off frame tracking for the MOVE gate (ADVANCED).
+
+  * :meth:`IQMClient.submit_circuit` now takes a :class:`CircuitCompilationOptions` parameter instead of ``max_circuit_duration_over_t2`` and ``heralding_mode``.
+  * Moved the existing ``max_circuit_duration_over_t2`` parameter to :class:`CircuitCompilationOptions`.
+  * Moved the existing ``heralding_mode`` parameter to :class:`CircuitCompilationOptions`.
+  * Introduced new option ``move_gate_validation`` to turn off MOVE gate validation during compilation (ADVANCED).
+  * Introduced new option ``move_gate_frame_tracking`` to turn off frame tracking for the MOVE gate (ADVANCED).
+  * New options can only be used on stations with ``CoCoS`` version 29.9 or later that support the MOVE gate instruction. Otherwise, the options will be ignored.
 
 Version 17.8
 ============